PC_HOST: add overridelock argument
Add --overridelock argument, which causes daft to ignore all lockfiles. This is useful when testing devices manually or troubleshooting, because using this argument we can lock the file (so no one else uses it) and call daft directly. Currently to achieve the same thing, one needs to manually SSH to one of the Beaglebones and run aft instead.

Signed-off-by: Christian da Costa <christian.da.costa@intel.com>